补充：Server 模块存储引擎的实现

+++ raft/src/main/java/zbl/moonlight/raft/log/RaftLogEntry.java
@@ -8,7 +8,7 @@ import java.nio.channels.SelectionKey;
 
 public record RaftLogEntry (
         SelectionKey selectionKey,
-        int serial,
+        long serial,
         int term,
         byte type,
         byte[] command
